+++ R/tardis.R
@@ -43,6 +43,10 @@
 #'                       of sentiment-bearing terms in ALL CAPS. Should probably
 #'                       be more than 1, to increase effects.
 #' @param use_punctuation Boolean, default TRUE. Should we consider sentence-level punctuation?
+#' @param simple_count Boolean, default FALSE. Convenience parameter that overrides many
+#'                     other parameters to enable simple counts of dictionary words:
+#'                     no modifiers, negations, capitalization, or punctuation
+#'                     effects are considered and no sigmoid function is applied.
 #' @param verbose For debugging--should it print lots of messages to the console?
 #'
 #' @return A `tbl_df` with one row for each input text and three new columns:
@@ -60,10 +64,21 @@ tardis <- function(
   negation_factor = 0.75,
   allcaps_factor = 1.25,
   use_punctuation = TRUE,
+  simple_count = FALSE,
   verbose = FALSE
 ) {
   # for dplyr data masking
   sentences_orig <- sentence <- word <- negation1 <- negation2 <- negation3 <- modifier1 <- modifier2 <- modifier3 <- text_id <- sentence_id <- sentiment_word <- punct_exclamation <- punct_question <- sentence_sum <- sentence_punct <- sentence_score <- NULL
+
+  if (simple_count) {
+    warning("Parameter simple_count = TRUE overrides most other parameters. Make sure this is intended!")
+    dict_modifiers <- "none"
+    dict_negations <- "none"
+    sigmoid_factor <- NA
+    negation_factor <- 1
+    allcaps_factor <- 1
+    use_punctuation <- FALSE
+  }
 
   # multiplicative scale factors for negations and all caps words
   # allcaps_factor needs 1 subtracted here because of how it's treated later